Revert lti_consumer/lti.py to old code; move new consumer under lti_consumer/lti_1p1/ and fix tests/imports
