+++ lti_consumer/__init__.py
@@ -1,4 +1,4 @@
 """
 Runtime will load the XBlock class from here.
 """
-from .lti_consumer_xblock import LtiConsumerXBlock
+from .lti_consumer import LtiConsumerXBlock
