+++ lti_consumer/tests/unit/test_utils.py
@@ -61,7 +61,7 @@ def patch_signed_parameters(func):
     Prepare the patches for the get_signed_lti_parameters function for tests.
     """
     func = patch(
-        'lti_consumer.lti_consumer_base.get_oauth_request_signature',
+        'lti_consumer.lti.get_oauth_request_signature',
         Mock(return_value=(
             'OAuth oauth_nonce="fake_nonce", '
             'oauth_timestamp="fake_timestamp", oauth_version="fake_version", oauth_signature_method="fake_method", '
@@ -70,17 +70,17 @@ def patch_signed_parameters(func):
     )(func)
 
     func = patch(
-        'lti_consumer.lti_consumer_xblock.LtiConsumerXBlock.prefixed_custom_parameters',
+        'lti_consumer.lti_consumer.LtiConsumerXBlock.prefixed_custom_parameters',
         PropertyMock(return_value={u'custom_param_1': 'custom1', u'custom_param_2': 'custom2'})
     )(func)
 
     func = patch(
-        'lti_consumer.lti_consumer_xblock.LtiConsumerXBlock.lti_provider_key_secret',
+        'lti_consumer.lti_consumer.LtiConsumerXBlock.lti_provider_key_secret',
         PropertyMock(return_value=('t', 's'))
     )(func)
 
     func = patch(
-        'lti_consumer.lti_consumer_xblock.LtiConsumerXBlock.user_id', PropertyMock(return_value=FAKE_USER_ID)
+        'lti_consumer.lti_consumer.LtiConsumerXBlock.user_id', PropertyMock(return_value=FAKE_USER_ID)
     )(func)
 
     return func
